Finished moving the peoplesearch module underneath the "./src/modules" folder.
