+++ client/src/modules/peoplesearch/main.dev.ts
@@ -22,12 +22,12 @@
 
 
 import { bootstrap, module } from 'angular';
-import ConfigService from './services/peoplesearch-config.service.dev';
-import peopleSearchModule from './modules/peoplesearch/peoplesearch.module';
-import PeopleService from './services/people.service.dev';
-import PwmService from './services/pwm.service.dev';
-import routes from './routes';
-import routeErrorHandler from './route-error-handler';
+import ConfigService from '../../services/peoplesearch-config.service.dev';
+import peopleSearchModule from './peoplesearch.module';
+import PeopleService from '../../services/people.service.dev';
+import PwmService from '../../services/pwm.service.dev';
+import routes from '../../routes';
+import routeErrorHandler from '../../route-error-handler';
 import uiRouter from '@uirouter/angularjs';
