+++ client/tsconfig.json
@@ -13,7 +13,7 @@
         "target": "es5"
     },
     "files": [
-        "src/main.ts"
+        "src/modules/peoplesearch/main.ts"
     ],
     "exclude": [
         "dist",
