+++ client/webpack.dev.js
@@ -28,7 +28,7 @@ var webpackMerge = require('webpack-merge');
 module.exports = webpackMerge(commonConfig, {
     devtool: 'cheap-module-source-map',
     entry: {
-        'peoplesearch.ng': './src/main.dev',
+        'peoplesearch.ng': './src/modules/peoplesearch/main.dev',
         'changepassword.ng': './src/modules/changepassword/changepassword.module',
         'configeditor.ng': './src/modules/configeditor/configeditor.module',
         'helpdesk.ng': './src/modules/helpdesk/main.dev'
